pyproject.toml
[project]
name = "zenoh-ros-type"
version = "0.3.0"
description = "Common class for ROS 2 message used by Zenoh"
authors = [
    {name = "ChenYing Kuo", email = "evshary@gmail.com"}
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pycdr2>=1.0,<2.0",
]

[project.urls]
repository = "https://github.com/evshary/zenoh-ros-type-python"
Issues = "https://github.com/evshary/zenoh-ros-type-python/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# We only want to build the zenoh_ros_type package
[tool.setuptools]
packages = ["zenoh_ros_type"]

# Only for examples
[project.optional-dependencies]
dev = [
    "eclipse-zenoh>=1.7.2,<2.0",
]

# Only for developer tests
[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-timeout>=2.4.0",
]

# Ruff
[tool.ruff]
line-length = 150

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "single"
